Moved tests from BasicIT to QuayGitHubBasicIT and QuayGitlabBasicIT
