+++ dockstore-cli-integration-testing/src/test/java/io/dockstore/client/cli/QuayGitlabBasicIT.java
@@ -1,0 +1,126 @@
+package io.dockstore.client.cli;
+
+import io.dockstore.common.CommonTestUtilities;
+import io.dockstore.common.ConfidentialTest;
+import io.dockstore.common.FlushingSystemErrRule;
+import io.dockstore.common.FlushingSystemOutRule;
+import io.dockstore.common.Registry;
+import io.dockstore.common.SlowTest;
+import io.dockstore.common.ToolTest;
+import io.dropwizard.testing.ResourceHelpers;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.contrib.java.lang.system.ExpectedSystemExit;
+import org.junit.contrib.java.lang.system.SystemErrRule;
+import org.junit.contrib.java.lang.system.SystemOutRule;
+import org.junit.experimental.categories.Category;
+
+@Category({ ConfidentialTest.class, ToolTest.class })
+public class QuayGitlabBasicIT extends BaseIT {
+    @Rule
+    public final ExpectedSystemExit systemExit = ExpectedSystemExit.none();
+
+    @Rule
+    public final SystemOutRule systemOutRule = new FlushingSystemOutRule().enableLog().muteForSuccessfulTests();
+
+    @Rule
+    public final SystemErrRule systemErrRule = new FlushingSystemErrRule().enableLog().muteForSuccessfulTests();
+
+    @Before
+    @Override
+    public void resetDBBetweenTests() throws Exception {
+        CommonTestUtilities.cleanStatePrivate1(SUPPORT, testingPostgres);
+    }
+
+    /*
+     * Test Quay and Gitlab -
+     * These tests are focused on testing entries created from Quay and Gitlab repositories
+     */
+
+    /**
+     * Checks that the two Quay/Gitlab entries were automatically found
+     */
+    @Test
+    public void testQuayGitlabAutoRegistration() {
+        // Need to add these to the db dump (db dump 1)
+
+        final long count = testingPostgres.runSelectStatement(
+            "select count(*) from tool where  registry = '" + Registry.QUAY_IO.getDockerPath() + "' and giturl like 'git@gitlab.com%'",
+            long.class);
+        Assert.assertEquals("there should be 2 registered from Quay and Gitlab", 2, count);
+    }
+
+    /**
+     * Ensures that you can't publish an automatically added Quay/Gitlab entry with an alternate structure unless you change the Dockerfile and Dockstore.cwl locations
+     */
+    @Test
+    public void testQuayGitlabPublishAlternateStructure() {
+        systemExit.expectSystemExitWithStatus(Client.API_ERROR);
+        Client.main(new String[] { "--config", ResourceHelpers.resourceFilePath("config_file.txt"), "tool", "publish", "--entry",
+            "quay.io/dockstoretestuser/quayandgitlabalternate", "--script" });
+    }
+
+    /**
+     * Checks that you can properly publish and unpublish a Quay/Gitlab entry
+     */
+    @Test
+    public void testQuayAndGitlabPublishAndUnpublishAnentry() {
+        // Publish
+        Client.main(new String[] { "--config", ResourceHelpers.resourceFilePath("config_file.txt"), "tool", "publish", "--entry",
+            "quay.io/dockstoretestuser/quayandgitlab", "--script" });
+
+        final long count = testingPostgres
+            .runSelectStatement("select count(*) from tool where name = 'quayandgitlab' and ispublished='t'", long.class);
+        Assert.assertEquals("there should be 1 registered", 1, count);
+
+        // Unpublish
+        Client.main(new String[] { "--config", ResourceHelpers.resourceFilePath("config_file.txt"), "tool", "publish", "--unpub", "--entry",
+            "quay.io/dockstoretestuser/quayandgitlab", "--script" });
+
+        final long count2 = testingPostgres
+            .runSelectStatement("select count(*) from tool where name = 'quayandgitlab' and ispublished='t'", long.class);
+        Assert.assertEquals("there should be 0 registered", 0, count2);
+    }
+
+    /**
+     * Checks that you can manually publish and unpublish a Quay/Gitlab entry with an alternate structure, if the CWL and Dockerfile paths are defined properly
+     */
+    @Test
+    @Category(SlowTest.class)
+    public void testQuayGitlabManualPublishAndUnpublishAlternateStructure() {
+        // Manual Publish
+        Client.main(new String[] { "--config", ResourceHelpers.resourceFilePath("config_file.txt"), "tool", "manual_publish", "--registry",
+            Registry.QUAY_IO.name(), Registry.QUAY_IO.toString(), "--namespace", "dockstoretestuser", "--name", "quayandgitlabalternate",
+            "--git-url", "git@gitlab.com:dockstore.test.user/quayandgitlabalternate.git", "--git-reference", "master", "--toolname",
+            "alternate", "--cwl-path", "/testDir/Dockstore.cwl", "--dockerfile-path", "/testDir/Dockerfile", "--script" });
+
+        final long count = testingPostgres
+            .runSelectStatement("select count(*) from tool where toolname = 'alternate' and ispublished='t'", long.class);
+
+        Assert.assertEquals("there should be 1 entries, there are " + count, 1, count);
+
+        // Unpublish
+        Client.main(new String[] { "--config", ResourceHelpers.resourceFilePath("config_file.txt"), "tool", "publish", "--unpub", "--entry",
+            "quay.io/dockstoretestuser/quayandgitlabalternate/alternate", "--script" });
+        final long count2 = testingPostgres
+            .runSelectStatement("select count(*) from tool where toolname = 'alternate' and ispublished='t'", long.class);
+
+        Assert.assertEquals("there should be 0 entries, there are " + count2, 0, count2);
+
+    }
+
+    /**
+     * Ensures that one cannot register an existing Quay/Gitlab entry if you don't give it an alternate toolname
+     */
+    @Test
+    public void testQuayGitlabManuallyRegisterDuplicate() {
+        systemExit.expectSystemExitWithStatus(Client.API_ERROR);
+        Client.main(new String[] { "--config", ResourceHelpers.resourceFilePath("config_file.txt"), "tool", "manual_publish", "--registry",
+            Registry.QUAY_IO.name(), Registry.QUAY_IO.toString(), "--namespace", "dockstoretestuser", "--name", "quayandgitlab",
+            "--git-url", "git@gitlab.com:DockstoreTestUser/dockstore-whalesay.git", "--git-reference", "master", "--script" });
+    }
+
+
+}
